sessao10/requirements-fastapi.txt
# Requisitos para o exercício FastAPI de segurança
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5  # Necessário para processar form-data
email-validator>=1.1.3   # Para validação de emails
